--- a/pkg/logql/log/parser.go
+++ b/pkg/logql/log/parser.go
@@ -38,13 +38,15 @@ var (
 	errUnexpectedJSONObject = fmt.Errorf("expecting json object(%d), but it is not", jsoniter.ObjectValue)
 	errMissingCapture       = errors.New("at least one named capture must be supplied")
 	errFoundAllLabels       = errors.New("found all required labels")
+	errLabelDoesNotMatch    = errors.New("found a label with a matcher that didn't match")
 )
 
 type JSONParser struct {
 	prefixBuffer []byte // buffer used to build json keys
 	lbs          *LabelsBuilder
 
-	keys internedStringSet
+	keys        internedStringSet
+	parserHints ParserHint
 }
 
 // NewJSONParser creates a log stage that can parse a json log line and add properties as labels.
@@ -56,13 +58,15 @@ func NewJSONParser() *JSONParser {
 }
 
 func (j *JSONParser) Process(_ int64, line []byte, lbs *LabelsBuilder) ([]byte, bool) {
-	if lbs.ParserLabelHints().NoLabels() {
+	parserHints := lbs.ParserLabelHints()
+	if parserHints.NoLabels() {
 		return line, true
 	}
 
 	// reset the state.
 	j.prefixBuffer = j.prefixBuffer[:0]
 	j.lbs = lbs
+	j.parserHints = parserHints
 
 	if err := jsonparser.ObjectEach(line, j.parseObject); err != nil {
 		if errors.Is(err, errFoundAllLabels) {
@@ -70,23 +74,25 @@ func (j *JSONParser) Process(_ int64, line []byte, lbs *LabelsBuilder) ([]byte, 
 			return line, true
 		}
 
-		lbs.SetErr(errJSON)
-		lbs.SetErrorDetails(err.Error())
-		if lbs.ParserLabelHints().PreserveError() {
-			lbs.Set(logqlmodel.PreserveErrorLabel, "true")
+		if errors.Is(err, errLabelDoesNotMatch) {
+			// one of the label matchers does not match. The whole line can be thrown away
+			return line, false
 		}
+
+		addErrLabel(errJSON, err, lbs)
+
 		return line, true
 	}
 	return line, true
 }
 
 func (j *JSONParser) parseObject(key, value []byte, dataType jsonparser.ValueType, offset int) error {
+	var err error
 	switch dataType {
 	case jsonparser.String, jsonparser.Number, jsonparser.Boolean:
-		j.parseLabelValue(key, value, dataType)
+		err = j.parseLabelValue(key, value, dataType)
 	case jsonparser.Object:
 		prefixLen := len(j.prefixBuffer)
-		var err error
 		if ok := j.nextKeyPrefix(key); ok {
 			err = jsonparser.ObjectEach(value, j.parseObject)
 		}
@@ -95,13 +101,13 @@ func (j *JSONParser) parseObject(key, value []byte, dataType jsonparser.ValueTyp
 		return err
 	}
 
-	if j.lbs.ParserLabelHints().AllRequiredExtracted() {
+	if j.parserHints.AllRequiredExtracted() {
 		// Not actually an error. Parsing can be short-circuited
 		// and this tells jsonparser to stop parsing
 		return errFoundAllLabels
 	}
 
-	return nil
+	return err
 }
 
 // nextKeyPrefix load the next prefix in the buffer and tells if it should be processed based on hints.
@@ -114,7 +120,7 @@ func (j *JSONParser) nextKeyPrefix(key []byte) bool {
 	return j.lbs.ParserLabelHints().ShouldExtractPrefix(unsafeGetString(j.prefixBuffer))
 }
 
-func (j *JSONParser) parseLabelValue(key, value []byte, dataType jsonparser.ValueType) {
+func (j *JSONParser) parseLabelValue(key, value []byte, dataType jsonparser.ValueType) error {
 	// the first time we use the field as label key.
 	if len(j.prefixBuffer) == 0 {
 		key, ok := j.keys.Get(key, func() (string, bool) {
@@ -128,10 +134,13 @@ func (j *JSONParser) parseLabelValue(key, value []byte, dataType jsonparser.Valu
 			return field, true
 		})
 		if !ok {
-			return
+			return nil
 		}
 		j.lbs.Set(key, readValue(value, dataType))
-		return
+		if !j.parserHints.ShouldContinueParsingLine(key, j.lbs) {
+			return errLabelDoesNotMatch
+		}
+		return nil
 
 	}
 	// otherwise we build the label key using the buffer
@@ -144,7 +153,7 @@ func (j *JSONParser) parseLabelValue(key, value []byte, dataType jsonparser.Valu
 		if j.lbs.BaseHas(string(j.prefixBuffer)) {
 			j.prefixBuffer = append(j.prefixBuffer, duplicateSuffix...)
 		}
-		if !j.lbs.ParserLabelHints().ShouldExtract(string(j.prefixBuffer)) {
+		if !j.parserHints.ShouldExtract(string(j.prefixBuffer)) {
 			return "", false
 		}
 
@@ -154,9 +163,14 @@ func (j *JSONParser) parseLabelValue(key, value []byte, dataType jsonparser.Valu
 	// reset the prefix position
 	j.prefixBuffer = j.prefixBuffer[:prefixLen]
 	if !ok {
-		return
+		return nil
 	}
+
 	j.lbs.Set(keyString, readValue(value, dataType))
+	if !j.parserHints.ShouldContinueParsingLine(keyString, j.lbs) {
+		return errLabelDoesNotMatch
+	}
+	return nil
 }
 
 func (j *JSONParser) RequiredLabelNames() []string { return []string{} }
@@ -237,6 +251,7 @@ func NewRegexpParser(re string) (*RegexpParser, error) {
 }
 
 func (r *RegexpParser) Process(_ int64, line []byte, lbs *LabelsBuilder) ([]byte, bool) {
+	parserHints := lbs.ParserLabelHints()
 	for i, value := range r.regex.FindSubmatch(line) {
 		if name, ok := r.nameIndex[i]; ok {
 			key, ok := r.keys.Get(unsafeGetBytes(name), func() (string, bool) {
@@ -247,7 +262,7 @@ func (r *RegexpParser) Process(_ int64, line []byte, lbs *LabelsBuilder) ([]byte
 				if lbs.BaseHas(sanitize) {
 					sanitize = fmt.Sprintf("%s%s", sanitize, duplicateSuffix)
 				}
-				if !lbs.ParserLabelHints().ShouldExtract(sanitize) {
+				if !parserHints.ShouldExtract(sanitize) {
 					return "", false
 				}
 
@@ -256,7 +271,11 @@ func (r *RegexpParser) Process(_ int64, line []byte, lbs *LabelsBuilder) ([]byte
 			if !ok {
 				continue
 			}
+
 			lbs.Set(key, string(value))
+			if !parserHints.ShouldContinueParsingLine(key, lbs) {
+				return line, false
+			}
 		}
 	}
 	return line, true
@@ -279,7 +298,8 @@ func NewLogfmtParser() *LogfmtParser {
 }
 
 func (l *LogfmtParser) Process(_ int64, line []byte, lbs *LabelsBuilder) ([]byte, bool) {
-	if lbs.ParserLabelHints().NoLabels() {
+	parserHints := lbs.ParserLabelHints()
+	if parserHints.NoLabels() {
 		return line, true
 	}
 
@@ -295,7 +315,7 @@ func (l *LogfmtParser) Process(_ int64, line []byte, lbs *LabelsBuilder) ([]byte
 				sanitized = fmt.Sprintf("%s%s", sanitized, duplicateSuffix)
 			}
 
-			if !lbs.ParserLabelHints().ShouldExtract(sanitized) {
+			if !parserHints.ShouldExtract(sanitized) {
 				return "", false
 			}
 			return sanitized, true
@@ -308,17 +328,21 @@ func (l *LogfmtParser) Process(_ int64, line []byte, lbs *LabelsBuilder) ([]byte
 		if bytes.ContainsRune(val, utf8.RuneError) {
 			val = nil
 		}
-		lbs.Set(key, string(val))
 
-		if lbs.ParserLabelHints().AllRequiredExtracted() {
+		lbs.Set(key, string(val))
+		if !parserHints.ShouldContinueParsingLine(key, lbs) {
+			return line, false
+		}
+
+		if parserHints.AllRequiredExtracted() {
 			break
 		}
 	}
 	if l.dec.Err() != nil {
-		lbs.SetErr(errLogfmt)
-		lbs.SetErrorDetails(l.dec.Err().Error())
-		if lbs.ParserLabelHints().PreserveError() {
-			lbs.Set(logqlmodel.PreserveErrorLabel, "true")
+		addErrLabel(errLogfmt, l.dec.Err(), lbs)
+
+		if !parserHints.ShouldContinueParsingLine(logqlmodel.ErrorLabel, lbs) {
+			return line, false
 		}
 		return line, true
 	}
@@ -349,7 +373,8 @@ func NewPatternParser(pn string) (*PatternParser, error) {
 }
 
 func (l *PatternParser) Process(_ int64, line []byte, lbs *LabelsBuilder) ([]byte, bool) {
-	if lbs.ParserLabelHints().NoLabels() {
+	parserHints := lbs.ParserLabelHints()
+	if parserHints.NoLabels() {
 		return line, true
 	}
 	matches := l.matcher.Matches(line)
@@ -360,11 +385,14 @@ func (l *PatternParser) Process(_ int64, line []byte, lbs *LabelsBuilder) ([]byt
 			name = name + duplicateSuffix
 		}
 
-		if !lbs.parserKeyHints.ShouldExtract(name) {
+		if !parserHints.ShouldExtract(name) {
 			continue
 		}
 
 		lbs.Set(name, string(m))
+		if !parserHints.ShouldContinueParsingLine(name, lbs) {
+			return line, false
+		}
 	}
 	return line, true
 }
@@ -535,20 +563,14 @@ func (j *JSONExpressionParser) Process(_ int64, line []byte, lbs *LabelsBuilder)
 	// the parser will pass an error if other
 	// parts of the line are malformed
 	if !isValidJSONStart(line) {
-		lbs.SetErr(errJSON)
-		if lbs.ParserLabelHints().PreserveError() {
-			lbs.Set(logqlmodel.PreserveErrorLabel, "true")
-		}
+		addErrLabel(errJSON, nil, lbs)
 		return line, true
 	}
 
 	var matches int
 	jsonparser.EachKey(line, func(idx int, data []byte, typ jsonparser.ValueType, err error) {
 		if err != nil {
-			lbs.SetErr(errJSON)
-			if lbs.ParserLabelHints().PreserveError() {
-				lbs.Set(logqlmodel.PreserveErrorLabel, "true")
-			}
+			addErrLabel(errJSON, err, lbs)
 			return
 		}
 
@@ -617,25 +639,38 @@ func (u *UnpackParser) Process(_ int64, line []byte, lbs *LabelsBuilder) ([]byte
 		return line, true
 	}
 
+	// we only care about object and values.
+	if line[0] != '{' {
+		addErrLabel(errJSON, errUnexpectedJSONObject, lbs)
+		return line, true
+	}
+
 	u.lbsBuffer = u.lbsBuffer[:0]
 	entry, err := u.unpack(line, lbs)
 	if err != nil {
-		lbs.SetErr(errJSON)
-		lbs.SetErrorDetails(err.Error())
-		if lbs.ParserLabelHints().PreserveError() {
-			lbs.Set(logqlmodel.PreserveErrorLabel, "true")
+		if errors.Is(err, errLabelDoesNotMatch) {
+			return entry, false
 		}
+		addErrLabel(errJSON, err, lbs)
 		return line, true
 	}
+
 	return entry, true
 }
 
-func (u *UnpackParser) unpack(entry []byte, lbs *LabelsBuilder) ([]byte, error) {
-	// we only care about object and values.
-	if entry[0] != '{' {
-		return nil, errUnexpectedJSONObject
+func addErrLabel(msg string, err error, lbs *LabelsBuilder) {
+	lbs.SetErr(msg)
+
+	if err != nil {
+		lbs.SetErrorDetails(err.Error())
 	}
 
+	if lbs.ParserLabelHints().PreserveError() {
+		lbs.Set(logqlmodel.PreserveErrorLabel, "true")
+	}
+}
+
+func (u *UnpackParser) unpack(entry []byte, lbs *LabelsBuilder) ([]byte, error) {
 	var isPacked bool
 	err := jsonparser.ObjectEach(entry, func(key, value []byte, typ jsonparser.ValueType, _ int) error {
 		switch typ {
@@ -683,6 +718,9 @@ func (u *UnpackParser) unpack(entry []byte, lbs *LabelsBuilder) ([]byte, error) 
 	if isPacked {
 		for i := 0; i < len(u.lbsBuffer); i = i + 2 {
 			lbs.Set(u.lbsBuffer[i], u.lbsBuffer[i+1])
+			if !lbs.ParserLabelHints().ShouldContinueParsingLine(u.lbsBuffer[i], lbs) {
+				return entry, errLabelDoesNotMatch
+			}
 		}
 	}
 	return entry, nil