This PR moves Line Filters to be as early in a log pipeline as possible. 

For example:
`{app="foo"} | logfmt |="some stuff"` becomes `{app="foo"} |="some stuff" | logfmt`

Any LineFilter after a `LineFormat` stage will be moved to directly after the nearest `LineFormat` stage

benchmarks:
```
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/pkg/logql/syntax
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
                    │ reorder_old.txt │           reorder_new.txt           │
                    │     sec/op      │   sec/op     vs base                │
ReorderedPipeline-8      2104.5n ± 3%   173.9n ± 2%  -91.74% (p=0.000 n=10)

                    │ reorder_old.txt │          reorder_new.txt          │
                    │      B/op       │   B/op    vs base                 │
ReorderedPipeline-8        336.0 ± 0%   0.0 ± 0%  -100.00% (p=0.000 n=10)

                    │ reorder_old.txt │          reorder_new.txt           │
                    │    allocs/op    │ allocs/op  vs base                 │
ReorderedPipeline-8        16.00 ± 0%   0.00 ± 0%  -100.00% (p=0.000 n=10)
```